+++ package.json
@@ -28,7 +28,7 @@
   "author": "Grafana Labs <team@grafana.com> (https://grafana.com)",
   "license": "Apache-2.0",
   "dependencies": {
-    "@types/diff": "^5.0.2",
+    "@types/node-fetch": "^2.6.2",
     "chalk": "^4.1.2",
     "debug": "^4.3.2",
     "diff": "^5.1.0",
@@ -43,8 +43,12 @@
   },
   "devDependencies": {
     "@grafana/eslint-config": "^4.0.0",
+    "@types/debug": "^4.1.7",
+    "@types/diff": "^5.0.2",
     "@types/jest": "^27.4.0",
     "@types/node": "^17.0.10",
+    "@types/rimraf": "^3.0.2",
+    "@types/tar": "^6.1.2",
     "@types/yargs": "^17.0.7",
     "@typescript-eslint/eslint-plugin": "^5.30.7",
     "@typescript-eslint/parser": "^5.30.7",
