+++ src/index.ts
@@ -1,9 +1,10 @@
 export * from './types';
-export * from './utils.compare';
-export * from './utils.compiler.imports';
-export * from './utils.compiler.imports';
-export * from './utils.compiler';
-export * from './utils.diff';
-export * from './utils.file';
-export * from './utils.npm';
-export * from './utils.compare.source';
+export * from './commands/compare';
+export * from './compiler/imports';
+export * from './compiler/imports';
+export * from './utils/typescript';
+export * from './utils/diff';
+export * from './utils/file';
+export * from './utils/npm';
+export * from './comparison/source';
+export * from './print';
