+++ src/print/imports.ts
@@ -1,6 +1,5 @@
 import chalk from 'chalk';
-import { ExportsInfo, ImportInfo } from './types';
-import { debug } from './utils.log';
+import { ImportInfo } from '../types';
 
 export function printImports({
   imports,
@@ -50,28 +49,4 @@ export function printImports({
         });
     });
   }
-}
-
-export function printExports(exports: ExportsInfo) {
-  debug('Printing results...');
-
-  console.log('');
-  console.log('List of exported members:');
-  console.log('===================================');
-  console.log(Object.keys(exports.exports).forEach((name) => console.log(`\t - ${name}`)));
-  console.log('===================================');
-}
-
-export function indentLines(str: string, tabsCount: number) {
-  return str.split('\n').reduce((acc, line) => `${acc}${indentLine(line, tabsCount)}\n`, '');
-}
-
-export function indentLine(str: string, count: number) {
-  let tabs = '';
-
-  for (let i = 0; i < count; i++) {
-    tabs += '\t';
-  }
-
-  return `${tabs}${str}`;
 }
