+++ src/utils/file.ts
@@ -1,5 +1,4 @@
 import fs from 'fs';
-import { logError } from './utils.log';
 
 export function pathExists(path: string): boolean {
   try {
